Support for CLDF post-processing
Sometimes it is necessary to post-process the CLDF on a per-dictionary basis.
E.g. for daakaka, all media files must be augmented with a description, to
specify the copyright.

This is now possible. To opt into postprocessing, dictionaries must define a
`postprocess` function in their python module, taking the `pathlib.Path` to
the `processed` directory as sole argument.